+++ th2_data_services/events_tree/event_tree.py
@@ -11,27 +11,25 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-
 from typing import List, Tuple, Generator, Callable, Optional, Union
 
 from treelib import Tree, Node
-from treelib.exceptions import NodeIDAbsentError
+from treelib.exceptions import NodeIDAbsentError, LoopError
 
-from th2_data_services.events_tree.exceptions import EventIdNotInTree
+from th2_data_services.events_tree.exceptions import EventIdNotInTree, EventAlreadyExist, EventRootExist, TreeLoop
 
-Th2Event = dict
+Th2Event = dict  # TODO - move to types. Also this class knows that th2-event is a dict, but it cannot to know.
 
 
-class EventsTree:
-    """EventsTree is a tree-based data structure of events.
+class EventTree:
+    """EventTree is a tree-based data structure of events.
 
     - get_x methods raise Exceptions if no result is found.
     - find_x methods return None if no result is found.
-    - EventsTree stores events as Nodes and interacts with them using an internal tree.
-    - EventsTree removes the 'body' field by default to save memory, but you can keep it.
-    - Note that EventsTree stores only one tree.
-        If you want to store all trees, use EventsTreeCollections.
-    - EventsTree contains all events in memory.
+    - EventTree stores events as Nodes and interacts with them using an internal tree.
+    - Note that EventTree stores only one tree.
+        If you want to store all trees, use EventTreeCollections.
+    - EventTree contains all events in memory.
 
     Take a look at the following HTML tree to understand some important terms.
 
@@ -45,25 +43,52 @@ class EventsTree:
     ```
     """
 
-    def __init__(self, tree: Tree):
-        """EventsTree constructor.
+    def __init__(self, event_name: str, event_id: str, data: dict = None):
+        """EventTree constructor.
 
         Args:
-            tree (treelib.Tree): Tree.
+            event_name: Event Name.
+            event_id: Event Id.
+            data: Data of event.
         """
-        self._tree = tree
+        self._tree = Tree()
+        self._create_root_event(event_name, event_id, data)
 
-    def _append_node(self, node: Node, parent_id: str) -> None:
-        """Appends a node to the tree.
+    def _create_root_event(self, event_name: str, event_id: str, data: dict = None) -> None:
+        """Appends a root event to the tree.
 
         Args:
-            node: Node.
-            parent_id: Parent event id.
+            event_name: Event Name.
+            event_id: Event Id.
+            data: Data of event.
         """
-        if parent_id not in self._tree:
-            raise NodeIDAbsentError(f"Node {parent_id} is not in the tree.")
+        if self._tree.root is not None:
+            raise EventRootExist(event_id)
+        if event_id in self._tree:
+            raise EventAlreadyExist(event_id)
 
-        self._tree.add_node(node, parent_id)
+        self._tree.create_node(tag=event_name, identifier=event_id, parent=None, data=data)
+
+    def append_event(self, event_name: str, event_id: str, parent_id: str, data: dict = None) -> None:
+        """Appends the event to the tree.
+
+        Args:
+            event_name: Event Name.
+            event_id: Event Id.
+            parent_id: Parent Id.
+            data: Data of event.
+        """
+        if event_id in self._tree:
+            raise EventAlreadyExist(event_id)
+        if parent_id is None:
+            raise ValueError(
+                "The param 'parent_id' must be not None. If you want create root event then use 'create_root_event'"
+            )
+
+        try:
+            self._tree.create_node(tag=event_name, identifier=event_id, parent=parent_id, data=data)
+        except NodeIDAbsentError:
+            raise EventIdNotInTree(parent_id)
 
     def get_all_events_iter(self) -> Generator[Th2Event, None, None]:
         """Returns all events from the tree as iterator."""
@@ -88,9 +113,53 @@ class EventsTree:
             raise EventIdNotInTree(id)
         return node.data
 
-    # TODO: In future it will be added.
-    # def __getitem__(self, item):
-    #     pass
+    def __getitem__(self, id_: str) -> Th2Event:
+        """e.g. ET['id1'] returns event.data."""
+        try:
+            return self._tree[id_].data
+        except NodeIDAbsentError:
+            raise EventIdNotInTree(id_)
+
+    def __setitem__(self, id_: str, data: dict) -> None:
+        # TODO - It shouldn't raise an exception.
+        #   It should create new Node or change existing
+        try:
+            self._tree[id_].data = data
+        except NodeIDAbsentError:
+            raise EventIdNotInTree(id_)
+
+    def update_event_name(self, event_id: str, event_name: str) -> None:
+        """Updates Event name in the tree. Note that it doesn't change internal data.
+
+        Args:
+            event_id: Event id.
+            event_name: Event name.
+
+        Raises:
+            EventIdNotInTree: If event id is not in the tree.
+        """
+        try:
+            self._tree.update_node(event_id, tag=event_name)
+        except NodeIDAbsentError:
+            raise EventIdNotInTree(event_id)
+
+    def update_parent_link(self, event_id: str, parent_id: str) -> None:
+        """Updates the link to parent.
+
+        Args:
+            event_id: Event id.
+            parent_id: New parent id.
+
+        Raises:
+            EventIdNotInTree: If event id is not in the tree.
+            TreeLoop: If parent id will point to the descendant event.
+        """
+        try:
+            self._tree.move_node(event_id, parent_id)
+        except NodeIDAbsentError:
+            raise EventIdNotInTree(event_id)
+        except LoopError:
+            raise TreeLoop(event_id, parent_id)
 
     def get_root_id(self) -> str:
         """Returns the root id."""
@@ -346,7 +415,7 @@ class EventsTree:
                 return event
         return None
 
-    def get_subtree(self, id: str) -> "EventsTree":
+    def get_subtree(self, id: str) -> "EventTree":
         """Returns subtree of the event by its id.
 
         Args:
@@ -361,10 +430,29 @@ class EventsTree:
         subtree = self._tree.subtree(id)
         if not subtree:
             raise EventIdNotInTree(id)
-        return EventsTree(subtree)
+
+        et = EventTree(event_name="0", event_id="0")
+        et._tree = subtree
+
+        return et
+
+    def merge_tree(self, parent_id: str, other_tree: "EventTree", use_deepcopy: bool = False) -> None:
+        """Merges a EventTree to specified identifier.
+
+        Args:
+            parent_id: Event id to which merge.
+            other_tree: EventTree.
+            use_deepcopy: True if you need deepcopy for your objects in event.
+
+        Raises:
+             EventIdNotInTree: If event id is not in the tree.
+        """
+        if parent_id not in self._tree:
+            raise EventIdNotInTree(parent_id)
+        self._tree.merge(parent_id, other_tree._tree, use_deepcopy)
 
     def show(self) -> None:
-        """Prints the EventsTree as tree view.
+        """Prints the EventTree as tree view.
 
         For example:
 
@@ -379,6 +467,20 @@ class EventsTree:
             |    |___ C31
         ```
         """
+        # TODO
+        # et.append_event('a', '2', et.get_root_id())
+        # et.show()
+        # Traceback (most recent call last):
+        #   File "C:\Users\admin\AppData\Local\Programs\Python\Python39\lib\code.py", line 90, in runcode
+        #     exec(code, self.locals)
+        #   File "<input>", line 1, in <module>
+        #   File "C:\Users\admin\exactpro\prj\th2\internal\DS\github\th2-data-services\th2_data_services\events_tree\event_tree.py", line 475, in show
+        #     self._tree.show()
+        #   File "C:\Users\admin\exactpro\prj\th2\internal\DS\github\th2-data-services\ds_lib_venv_py39\lib\site-packages\treelib\tree.py", line 854, in show
+        #     print(self._reader)
+        #   File "C:\Users\admin\AppData\Local\Programs\Python\Python39\lib\encodings\cp1251.py", line 19, in encode
+        #     return codecs.charmap_encode(input,self.errors,encoding_table)[0]
+        # UnicodeEncodeError: 'charmap' codec can't encode characters in position 6-8: character maps to <undefined>
         self._tree.show()
 
     def __len__(self) -> int: